+++ marko.json
@@ -212,7 +212,7 @@
     "@no-text": "boolean"
   },
   "<ebay-filter>": {
-    "renderer": "./src/components/ebay-filter/index.js",
+    "renderer": "./src/components/ebay-filter/index.marko",
     "@*": {
       "targetProperty": null,
       "type": "expression"
