--- src/components/ebay-filter/template.marko
@@ -1,28 +1,0 @@
-<script marko-init>
-    var processHtmlAttributes = require("../../common/html-attributes");
-    var ignoredAttributes = ["class", "style", "disabled", "href", "priority", "selected", "a11ySelectedText"];
-</script>
-
-<var classPrefix=(data.href ? "filter-link" : "filter-button")/>
-
-<${data.href ? "a" : "button"}
-    w-bind
-    w-on-click="handleButtonClick"
-    class=[
-        classPrefix,
-        data.selected ? classPrefix + "--selected" : classPrefix + "--unselected",
-        data.class
-    ]
-    style=data.style
-    href=data.href
-    disabled=data.disabled
-    type=(!data.href && "button")
-    aria-pressed=(!data.href && data.selected && "true")
-    ${processHtmlAttributes(data, ignoredAttributes)}>
-    <span class="${classPrefix}__cell">
-        <span w-body/>
-        <if(data.href && data.selected)>
-            <span w-id="active-text" class="clipped"> - ${data.a11ySelectedText || "Selected"}</span>
-        </if>
-    </span>
-</>
